Core: Additional conversion and narrowing warning fixes for Clang #rb marc.audy [FYI] Steve.Robb #jira UE-87131
#ROBOMERGE-SOURCE: CL 11140226 via CL 11140233 via CL 11140235
#ROBOMERGE-BOT: (v640-11091645)

[CL 11140236 by michael noland in Main branch]

+++ Engine/Source/Runtime/Core/Public/Delegates/DelegateBase.h
@@ -256,5 +256,5 @@ private:
 
 inline void* operator new(size_t Size, FDelegateBase& Base)
 {
-	return Base.Allocate(Size);
+	return Base.Allocate((int32)Size);
 }
