+++ Engine/Source/Runtime/Core/Public/GenericPlatform/StandardPlatformString.h
@@ -62,7 +62,7 @@ public:
 
 	static FORCEINLINE int32 Strlen( const WIDECHAR* String )
 	{
-		return wcslen( String );
+		return (int32)wcslen( String );
 	}
 
 	static FORCEINLINE const WIDECHAR* Strstr( const WIDECHAR* String, const WIDECHAR* Find)
@@ -82,7 +82,7 @@ public:
 
 	static FORCEINLINE int32 Atoi(const WIDECHAR* String)
 	{
-		return wcstol( String, NULL, 10 );
+		return (int32)wcstol( String, NULL, 10 );
 	}
 
 	static FORCEINLINE int64 Atoi64(const WIDECHAR* String)
@@ -102,7 +102,7 @@ public:
 
 	static FORCEINLINE int32 Strtoi( const WIDECHAR* Start, WIDECHAR** End, int32 Base )
 	{
-		return wcstol( Start, End, Base );
+		return (int32)wcstol( Start, End, Base );
 	}
 
 	static FORCEINLINE int64 Strtoi64( const WIDECHAR* Start, WIDECHAR** End, int32 Base )
@@ -223,7 +223,7 @@ public:
 
 	static FORCEINLINE int32 Strlen( const ANSICHAR* String )
 	{
-		return strlen( String );
+		return (int32)strlen( String );
 	}
 
 	static FORCEINLINE const ANSICHAR* Strstr( const ANSICHAR* String, const ANSICHAR* Find)
@@ -263,7 +263,7 @@ public:
 
 	static FORCEINLINE int32 Strtoi( const ANSICHAR* Start, ANSICHAR** End, int32 Base )
 	{
-		return strtol( Start, End, Base );
+		return (int32)strtol( Start, End, Base );
 	}
 
 	static FORCEINLINE int64 Strtoi64( const ANSICHAR* Start, ANSICHAR** End, int32 Base )
